// sanity.types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export type Slug = {
  _type: "slug";
  current: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type USER_CART_QUERY_RESULT = {
  _id: string;
  _type: "cart";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  clerkUserId?: string;
  items: Array<{
    product: {
      _id: string;
      _type: "product";
      _createdAt: string;
      _updatedAt: string;
      _rev: string;
      name: string;
      slug: Slug;
      image?: {
        asset?: SanityImageAssetReference;
        media?: unknown;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        _type: "image";
      };
      description?: string;
      price: number;
      discount: number;
      categories:
        | Array<{
            _id: string;
            _type: "category";
            _createdAt: string;
            _updatedAt: string;
            _rev: string;
            title: string;
            slug: Slug;
            description?: string;
          }>
        | Array<never>;
      stock?: number;
      label?: string;
      status?: "hot" | "new" | "sale";
    };
    quantity: number;
    _key: string;
  }>;
} | null;

// Source: sanity/helpers/index.ts
// Variable: GUEST_CART_QUERY
// Query: *[_type == "cart" && _id == $localCartId][0]{      ...,"items": coalesce(			items[]{			...,product->{			...,"categories": coalesce(				categories[]->,				[])}			},			[])    	}
export type GUEST_CART_QUERY_RESULT = {
  _id: string;
  _type: "cart";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  clerkUserId?: string;
  items: Array<{
    product: {
      _id: string;
      _type: "product";
      _createdAt: string;
      _updatedAt: string;
      _rev: string;
      name: string;
      slug: Slug;
      image?: {
        asset?: SanityImageAssetReference;
        media?: unknown;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        _type: "image";
      };
      description?: string;
      price: number;
      discount: number;
      categories:
        | Array<{
            _id: string;
            _type: "category";
            _createdAt: string;
            _updatedAt: string;
            _rev: string;
            title: string;
            slug: Slug;
            description?: string;
          }>
        | Array<never>;
      stock?: number;
      label?: string;
      status?: "hot" | "new" | "sale";
    };
    quantity: number;
    _key: string;
  }>;
} | null;